Refactor: delete entity in dto, move PostJobLimitDto.java
