+++ src/main/java/com/spring/tming/domain/post/dto/request/PostCreateReq.java
@@ -1,5 +1,6 @@
 package com.spring.tming.domain.post.dto.request;
 
+import com.spring.tming.domain.post.dto.PostJobLimitDto;
 import com.spring.tming.domain.post.entity.JobLimit;
 import com.spring.tming.global.meta.Skill;
 import java.sql.Timestamp;
@@ -15,7 +16,7 @@ public class PostCreateReq {
     private String title;
     private String content;
     private Timestamp deadline;
-    private List<JobLimit> jobLimits;
+    private List<PostJobLimitDto> jobLimits;
     private List<Skill> skills;
 
     @Builder
@@ -23,7 +24,7 @@ public class PostCreateReq {
             String title,
             String content,
             Timestamp deadline,
-            List<JobLimit> jobLimits,
+            List<PostJobLimitDto> jobLimits,
             List<Skill> skills) {
         this.title = title;
         this.content = content;
