+++ src/main/java/com/spring/tming/domain/post/dto/request/PostUpdateReq.java
@@ -1,5 +1,6 @@
 package com.spring.tming.domain.post.dto.request;
 
+import com.spring.tming.domain.post.dto.PostJobLimitDto;
 import com.spring.tming.domain.post.entity.JobLimit;
 import com.spring.tming.global.meta.Skill;
 import java.sql.Timestamp;
@@ -16,7 +17,7 @@ public class PostUpdateReq {
     private String title;
     private String content;
     private Timestamp deadline;
-    private List<JobLimit> jobLimits;
+    private List<PostJobLimitDto> jobLimits;
     private List<Skill> skills;
 
     @Builder
@@ -25,7 +26,7 @@ public class PostUpdateReq {
             String title,
             String content,
             Timestamp deadline,
-            List<JobLimit> jobLimits,
+            List<PostJobLimitDto> jobLimits,
             List<Skill> skills) {
         this.postId = postId;
         this.title = title;
