+++ src/main/java/com/spring/tming/domain/post/service/PostServiceMapper.java
@@ -1,7 +1,7 @@
 package com.spring.tming.domain.post.service;
 
 import com.spring.tming.domain.post.dto.response.PostCreateRes;
-import com.spring.tming.domain.post.dto.response.PostJobLimitRes;
+import com.spring.tming.domain.post.dto.PostJobLimitDto;
 import com.spring.tming.domain.post.dto.response.PostReadRes;
 import com.spring.tming.domain.post.entity.JobLimit;
 import com.spring.tming.domain.post.entity.Post;
@@ -43,19 +43,19 @@ public interface PostServiceMapper {
     }
 
     @Mapping(source = "jobLimits", target = "jobLimits")
-    default List<PostJobLimitRes> toPostJobLimitRes(List<JobLimit> jobLimits) {
+    default List<PostJobLimitDto> toPostJobLimitRes(List<JobLimit> jobLimits) {
         if (CollectionUtils.isEmpty(jobLimits)) {
             return null;
         }
-        List<PostJobLimitRes> postJobLimitRes = new ArrayList<>();
+        List<PostJobLimitDto> postJobLimitRes = new ArrayList<>();
         jobLimits.forEach(
-                jobLimit -> postJobLimitRes.add(PostServiceMapper.INSTANCE.toPostjobLimitRes(jobLimit)));
+                jobLimit -> postJobLimitRes.add(PostServiceMapper.INSTANCE.toPostjobLimitDto(jobLimit)));
         return postJobLimitRes;
     }
 
     PostCreateRes toPostCreateRes(Post post);
 
-    PostJobLimitRes toPostjobLimitRes(JobLimit jobLimit);
+    PostJobLimitDto toPostjobLimitDto(JobLimit jobLimit);
 
     @Mapping(source = "deadline", target = "deadline")
     @Mapping(source = "post.user.username", target = "username")
